+++ src/main/java/com/whiteday/aiecolink/domain/station/controller/StationController.java
@@ -6,8 +6,8 @@ import com.whiteday.aiecolink.domain.station.model.StationSummaryRes;
 import com.whiteday.aiecolink.domain.station.model.request.StationRegisterReq;
 import com.whiteday.aiecolink.domain.station.model.request.StationUpdateReq;
 import com.whiteday.aiecolink.domain.station.service.StationService;
-import com.whiteday.aiecolink.member.security.CustomUserDetails;
-import com.whiteday.aiecolink.member.User;
+import com.whiteday.aiecolink.domain.member.security.CustomUserDetails;
+import com.whiteday.aiecolink.domain.member.model.entity.User;
 import lombok.RequiredArgsConstructor;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
