+++ src/main/java/com/whiteday/aiecolink/domain/station/controller/TestStationController.java
@@ -7,9 +7,9 @@ import com.whiteday.aiecolink.domain.station.model.request.StationUpdateReq;
 import com.whiteday.aiecolink.domain.station.service.StationService;
 import com.whiteday.aiecolink.global.error.CustomException;
 import com.whiteday.aiecolink.global.error.ErrorCode;
-import com.whiteday.aiecolink.member.Role;
-import com.whiteday.aiecolink.member.User;
-import com.whiteday.aiecolink.member.UserRepository;
+import com.whiteday.aiecolink.domain.member.model.Role;
+import com.whiteday.aiecolink.domain.member.model.entity.User;
+import com.whiteday.aiecolink.domain.member.repository.UserRepository;
 import lombok.RequiredArgsConstructor;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
