+++ src/main/java/com/whiteday/aiecolink/domain/station/repository/StationRepository.java
@@ -1,7 +1,7 @@
 package com.whiteday.aiecolink.domain.station.repository;
 
 import com.whiteday.aiecolink.domain.station.model.entity.Station;
-import com.whiteday.aiecolink.member.User;
+import com.whiteday.aiecolink.domain.member.model.entity.User;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.data.jpa.repository.Query;
 import org.springframework.stereotype.Repository;
